% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps1_3.R
\name{c_psurvival}
\alias{c_psurvival}
\title{Conditional probability of survival}
\usage{
c_psurvival(c_pdeath)
}
\arguments{
\item{c_pdeath}{Conditional probability of death}
}
\value{
1-conditional probability of death
}
\description{
Given probability of death given age group, calculates probability of survival given age group
}
