% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps1_3.R
\name{central_death_rate}
\alias{central_death_rate}
\title{Central Death Rate}
\usage{
central_death_rate(data, age, pop, deaths, ...)
}
\arguments{
\item{data}{The mortality dataset, includes an age grouping variable,}

\item{age}{The age grouping variable, must be categorical}

\item{pop}{Population of each age group, must be numeric}

\item{deaths}{The midyear number of deaths at each age group, must be numeric}

\item{...}{Other optional grouping variables (can be race, gender, etc.)}
}
\description{
Central Death Rate
}
