% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps1_3.R
\name{central_death_rt}
\alias{central_death_rt}
\title{Central Death Rate}
\usage{
central_death_rt(D, P)
}
\arguments{
\item{D}{deaths/age group}

\item{P}{population/age group}
}
\value{
central death rate
}
\description{
Calculates the death rates for age groups
}
\examples{
central_death_rt(10,100)
}
