% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps1_3.R
\name{conditional_death_prob}
\alias{conditional_death_prob}
\title{Conditional Probability of Death at Age x}
\usage{
conditional_death_prob(data, age, pop, deaths)
}
\arguments{
\item{data}{The mortality dataset, includes an age grouping variable}

\item{age}{The age grouping variable, must be cateogrical}

\item{pop}{Population of each age group, must be numeric}

\item{deaths}{The number of deaths at each age group, must be numeric}
}
\value{
Data frame that was input with an added column, ConditionalProbDeath.
}
\description{
Adds a new column called ConditionalProbDeath to the dataset that was input.
This column represents the probability of death given the age group for each age group.
In other words, the probability a person in a given age group will die before their next birthday.
}
\examples{
# This function will add a ConditionalProbDeath column to the dataset
conditional_death_prob(mortality2, "age_group", "population", "deaths")
}
