% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps1_3.R
\name{cp_death}
\alias{cp_death}
\title{Conditional probability of death}
\usage{
cp_death(df, age, deaths, pop)
}
\arguments{
\item{df}{Data frame}

\item{age}{Age groups - column of data frame}

\item{deaths}{Number of deaths in each age group - column of a data frame}

\item{pop}{Population of each age group - column of a data frame}
}
\value{
conditional probability of death (double)
}
\description{
Conditional probability of death
}
