% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps1_3.R
\name{cp_survival}
\alias{cp_survival}
\title{Conditional probability of survival}
\usage{
cp_survival(df, age, deaths, pop)
}
\arguments{
\item{df}{Data frame}

\item{age}{Age groups - column of data frame}

\item{deaths}{Number of deaths in each age group - column of a data frame}

\item{pop}{Population of each age group - column of a data frame}
}
\value{
conditional probability of survival (double)
}
\description{
Conditional probability of survival
}
