% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable_func.R
\name{create_lifetable}
\alias{create_lifetable}
\title{Create full life table}
\usage{
create_lifetable(df, age, pop, death)
}
\arguments{
\item{death}{}
}
\value{
lifetable data frame
}
\description{
Adds columns to existing data frame for all of the steps of calculating
the life expectancy rate for the given age groups
}
