% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps1_3.R
\name{death_rate}
\alias{death_rate}
\title{Central death rate}
\usage{
death_rate(df, age, deaths, pop)
}
\arguments{
\item{df}{Data frame}

\item{age}{Age groups - column of data frame}

\item{deaths}{Number of deaths in each age group - column of a data frame}

\item{pop}{Population of each age group - column of a data frame}
}
\value{
Death rate (double)
}
\description{
Central death rate
}
