% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{input_check}
\alias{input_check}
\title{Error Handling Function}
\usage{
input_check(data, age, pop, deaths)
}
\arguments{
\item{data}{data frame input in the upper function}

\item{age}{age string or character input in the upper function}

\item{pop}{pop string or character input in the upper function}

\item{deaths}{deaths string or character input in the upper function}
}
\value{
data frame with numeric pop and deaths columns
}
\description{
Checks inputs data, age, pop, and deaths to make sure they are valid.
}
