% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps1_3.R
\name{makes_lifetable}
\alias{makes_lifetable}
\title{Making the lifetable}
\usage{
makes_lifetable(df, age, deaths, pop)
}
\arguments{
\item{deaths}{}
}
\value{
lifetable
}
\description{
Making the lifetable
}
