% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R, R/process_datasets.R
\docType{data}
\name{mortality2}
\alias{mortality2}
\title{Age-wise mortality dataset}
\format{
A data frame with 85 row of 3 columns representing different age groups
with which we use to make a life table
\describe{
\item{age_group}{Categorical variable identifying each age group}
\item{deaths}{the mid-year number of deaths in each age group}
\item{population}{the US population of each age group}
}

A data frame with 13 row of 3 columns representing different age groups
with which we use to make a life table
\describe{
\item{age_group}{Categorical variable identifying each age group}
\item{deaths}{the mid-year number of deaths in each age group}
\item{population}{the US population of each age group}
}
}
\source{
\url{https://wonder.cdc.gov/ucd-icd10.html}

\url{https://wonder.cdc.gov/ucd-icd10.html}
}
\usage{
mortality2

mortality2
}
\description{
A data extract takes from the CDC Wonder database with single-year age groups.
This data is from the year 2016.

A data extract takes from the CDC Wonder database.
}
\keyword{datasets}
