% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps4-5.R
\name{num_suriving_to_age_x}
\alias{num_suriving_to_age_x}
\title{Number Surviving to Age X (lx)}
\usage{
num_suriving_to_age_x(data, age, pop, deaths, ...)
}
\arguments{
\item{...}{}
}
\value{
original data frame with added columns
}
\description{
Calculates number of people surviving in each age group
}
