% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps4-6.R
\name{number_to_survive}
\alias{number_to_survive}
\title{The Number of People to Survive to Age x}
\usage{
number_to_survive(data, age, pop, deaths)
}
\arguments{
\item{data}{The mortality dataset, includes an age grouping variable,}

\item{age}{The age grouping variable, must be categorical}

\item{pop}{Population of each age group, must be numeric}

\item{deaths}{The midyear number of deaths at each age group, must be numeric}
}
\value{
Dataset that was input with added columns:
ConditionalProbDeath, ConditionalProbLife, and NumberToSurvive.
}
\description{
Adds a new column called NumberToSurvive to the dataset that was input.
NumberToSurvive represents the number of people living at the beginning of the given age interval,
using an arbitrary 100,000 people for the first age group in the table.
}
\examples{
# This function will add the ConditionalProbDeath, ConditionalProbLife, and
# NumberToSurvive columns to the dataset
number_to_survive(mortality2, "age_group", "population", "deaths")
}
