% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps4-6.R
\name{person_years}
\alias{person_years}
\title{Person Years Lived at Age x}
\usage{
person_years(data, age, pop, deaths)
}
\arguments{
\item{data}{The mortality dataset, includes an age grouping variable,}

\item{age}{The age grouping variable, must be categorical}

\item{pop}{Population of each age group, must be numeric}

\item{deaths}{The midyear number of deaths at each age group, must be numeric}
}
\value{
Dataset that was input with the added columns:
ConditionalProbDeath, ConditionalProbLife, NumberToSurvive, PropToSurvive, PersonYears.
}
\description{
Adds a new column called PersonYears to the dataset that was input.
PersonYears represents the number of years lived at age x based on the number surviving to age x.
}
\examples{
# This function will add the ConditionalProbDeath, ConditionalProbLife,
# NumberToSurvive, PropToSurvive, and PersonYears columns to the dataset
person_years(mortality2, "age_group", "population", "deaths")
}
