% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps4-5.R
\name{prop_survive_to_age_x}
\alias{prop_survive_to_age_x}
\title{Proportion Surviving To Age X}
\usage{
prop_survive_to_age_x(data, age, pop, deaths, ...)
}
\arguments{
\item{...}{}
}
\description{
Calculates the proportion of the
}
