% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps4-6.R
\name{prop_to_survive}
\alias{prop_to_survive}
\title{Proportion to Survive to Age x}
\usage{
prop_to_survive(data, age, pop, deaths, ...)
}
\arguments{
\item{data}{The mortality dataset, includes an age grouping variable,}

\item{age}{The age grouping variable, must be categorical}

\item{pop}{Population of each age group, must be numeric}

\item{deaths}{The midyear number of deaths at each age group, must be numeric}

\item{...}{Other optional grouping variables (can be race, gender, etc.)}
}
\description{
Proportion to Survive to Age x
}
