% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps7-8.R
\name{total_years_lived}
\alias{total_years_lived}
\title{Total Years Lived From Age x}
\usage{
total_years_lived(data, age, pop, deaths)
}
\arguments{
\item{data}{The mortality dataset, includes an age grouping variable,}

\item{age}{The age grouping variable, must be categorical}

\item{pop}{Population of each age group, must be numeric}

\item{deaths}{The midyear number of deaths at each age group, must be numeric}
}
\value{
Dataset that was input with the added columns:
ConditionalProbDeath, ConditionalProbLife, NumberToSurvive, PersonYears, and TotalYears.
}
\description{
Adds a new column called TotalYears to the dataset that was input.
TotalYears is the number of years lived from age zero to age x.
}
\examples{
# This function will add the ConditionalProbDeath, ConditionalProbLife,
# NumberToSurvive, PropToSurvive, PersonYearsm and TotalYears columns to the
# dataset
total_years_lived(mortality2, "age_group", "population", "deaths")
}
