% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps1_3.R
\name{c_pdeath}
\alias{c_pdeath}
\title{Conditional probability of death}
\usage{
c_pdeath(P, D)
}
\arguments{
\item{P}{population/age group}

\item{D}{deaths/age group}
}
\value{
conditional probability of death
}
\description{
Calculates the probability of death given age group
}
