% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps1_3.R
\name{central_death_rate}
\alias{central_death_rate}
\title{Central Death Rate}
\usage{
central_death_rate(data, age, pop, deaths)
}
\arguments{
\item{data}{The mortality dataset, includes an age grouping variable,}

\item{age}{The age grouping variable, must be categorical}

\item{pop}{Population of each age group, must be numeric}

\item{deaths}{The midyear number of deaths at each age group, must be numeric}
}
\value{
Data frame that was input with an added CentralDeathRate column.
}
\description{
Adds a new column called CentralDeathRate to the dataset that was input.
This column represents the central death rate of each age group - deaths/population.
}
\examples{
# This function adds a CentralDeathRate column to the dataset
central_death_rate(mortality2, "age_group", "population", "deaths")
}
