% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps1_3.R
\name{conditional_life_prob}
\alias{conditional_life_prob}
\title{Conditional Probability of Survival at Age x}
\usage{
conditional_life_prob(data, age, pop, deaths, ...)
}
\arguments{
\item{data}{The mortality dataset, includes an age grouping variable}

\item{age}{The age grouping variable, must be cateogrical}

\item{pop}{Population of each age group, must be numeric}

\item{deaths}{The number of deaths at each age group, must be numeric}

\item{...}{Optional other categorical grouping variables (race, sex, etc.)}
}
\description{
Conditional Probability of Survival at Age x
}
