% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{mortality}
\alias{mortality}
\title{A sample mortality data}
\format{
A data frame with 85 rows of 3 columns representing the US population
at mulit-year different age groups with which we use to make a life table.
This data is from the year 2018
\describe{
\item{age_group}{Categorical variable identifying each age group}
\item{deaths}{the mid-year number of deaths in each age group}
\item{population}{the US population of each age group}
}
}
\source{
\url{https://wonder.cdc.gov}
}
\usage{
mortality
}
\description{
A data extract takes from the CDC Wonder database.
}
\keyword{datasets}
