% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{mortality2}
\alias{mortality2}
\title{A sample mortality data}
\format{
A data frame with 85 rows of 3 columns representing the deaths and US
population at each single-year age group with which we can use to make a life table.
This data is from the year 2016.

\describe{
\item{age_group}{Categorical variable identifying each age group}
\item{deaths}{the mid-year number of deaths in each age group}
\item{population}{the US population of each age group}
}
}
\source{
\url{https://wonder.cdc.gov/ucd-icd10.html}
}
\usage{
mortality2
}
\description{
A data extract taken from the CDC Wonder database.
}
\keyword{datasets}
