% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps4-6.R
\name{number_to_survive}
\alias{number_to_survive}
\title{The Number of People to Survive to Age x}
\usage{
number_to_survive(data, age, pop, deaths, ...)
}
\arguments{
\item{data}{The mortality dataset, includes an age grouping variable,}

\item{age}{The age grouping variable, must be categorical}

\item{pop}{Population of each age group, must be numeric}

\item{deaths}{The midyear number of deaths at each age group, must be numeric}

\item{...}{Other optional grouping variables (can be race, gender, etc.)}
}
\description{
The Number of People to Survive to Age x
}
