% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steps4-6.R
\name{prop_to_survive}
\alias{prop_to_survive}
\title{Proportion to Survive to Age x}
\usage{
prop_to_survive(data, age, pop, deaths)
}
\arguments{
\item{data}{The mortality dataset, includes an age grouping variable,}

\item{age}{The age grouping variable, must be categorical}

\item{pop}{Population of each age group, must be numeric}

\item{deaths}{The midyear number of deaths at each age group, must be numeric}
}
\value{
Data frame that was input with columns for steps up to proportion surviving to age x included.
That is, the original data with the following added columns:
ConditionalProbDeath, ConditionalProbLife, NumberToSurvive, PropToSurvive
}
\description{
Adds a new column called PropToSurvive to the dataset that was input.
PropToSurvive is the proportion surviving to age x
}
\examples{
# This function will add the ConditionalProbDeath, ConditionalProbLife,
# NumberToSbrvivem and PropToSurvive columns to the dataset
prop_to_survive(mortality2, "age_group", "population", "deaths")
}
